VfsProjectCopier
### What's done:
 * Started to use `VfsUtil#copyDirectory` as core of `ProjectCopier` - it fixes `File Cache Conflict`
