+++ plugin/src/main/kotlin/com/sanyavertolet/kotlinjspreview/JsPreviewRunLineMarkerContributor.kt
@@ -6,9 +6,13 @@ import com.intellij.psi.PsiElement
 import com.sanyavertolet.kotlinjspreview.utils.isValOfPropertyAnnotatedWith
 
 class JsPreviewRunLineMarkerContributor : RunLineMarkerContributor() {
-    override fun getInfo(element: PsiElement): Info? = if (element.isValOfPropertyAnnotatedWith("JsPreview")) {
+    override fun getInfo(element: PsiElement): Info? = if (element.isValOfPropertyAnnotatedWith(PREVIEW_ANNOTATION)) {
         Info(AllIcons.Actions.Execute, { "Run preview" }, BuildKotlinJsAction(element.parent))
     } else {
         null
+    }
+
+    companion object {
+        private const val PREVIEW_ANNOTATION = "JsPreview"
     }
 }
